requirements.txt
attrs==25.4.0
certifi==2025.11.12
gherkin-official==29.0.0
h11==0.16.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
outcome==1.3.0.post0
packaging==25.0
parse==1.20.2
parse_type==0.6.6
pluggy==1.6.0
Pygments==2.19.2
PySocks==1.7.1
pytest==9.0.1
pytest-bdd==8.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
PyYAML==6.0.3
selenium==4.38.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.5.0
websocket-client==1.9.0
wsproto==1.3.2
axe-selenium-python==2.1.6
webdriver-manager==4.0.2